Split monitoring module
